requirements.txt
apache-airflow-providers-snowflake
apache-airflow-providers-google==4.0.0
google-analytics-data==0.10.0
